+++ app/Http/Controllers/Backend/KeagenankapalController.php
@@ -131,11 +131,15 @@ class KeagenankapalController extends Controller
      * @param  \App\Models\keagenan_kapal  $keagenan_kapal
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $this->authorize('view', KeagenanKapal::findOrFail($id));
-        KeagenanKapal::destroy($id);
-        return redirect()->route('keagenan_kapal.index')->with('success', 'Data berhasil dihapus');
+        if ($request->delete == 'true') {
+            $this->authorize('view', KeagenanKapal::findOrFail($id));
+            KeagenanKapal::destroy($id);
+            return redirect()->route('keagenan_kapal.index')->with('success', 'Data berhasil dihapus');
+        }
+        alert()->error('Gagal', 'Data gagal dihapus');
+        return redirect()->route('keagenan_kapal.index');
     }
 
     public function createBerangkat()
