+++ app/Http/Controllers/Backend/TersusController.php
@@ -123,12 +123,16 @@ class TersusController extends Controller
      * @param  \App\Models\Tersus  $tersus
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $this->authorize('view', Tersus::findOrFail($id));
-        Tersus::destroy($id);
-        storeLog(null, "User " . auth()->user()->name . " menghapus data tersus");
-        return redirect()->route('tersus.index')->with('success', 'Data berhasil dihapus');
+        if ($request->delete == 'true') {
+            $this->authorize('view', Tersus::findOrFail($id));
+            Tersus::destroy($id);
+            storeLog(null, "User " . auth()->user()->name . " menghapus data tersus");
+            return redirect()->route('tersus.index')->with('success', 'Data berhasil dihapus');
+        }
+        alert()->error('Gagal', 'Data gagal dihapus');
+        return redirect()->route('tersus.index');
     }
 
     public function createBerangkat()
