+++ app/Http/Controllers/Backend/UserController.php
@@ -101,10 +101,14 @@ class UserController extends Controller
      * @param  \App\Models\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        User::destroy($id);
-        return redirect()->route('user.index')->with('success', 'Data berhasil dihapus');
+        if ($request->delete == 'true') {
+            User::destroy($id);
+            return redirect()->route('user.index')->with('success', 'Data berhasil dihapus');
+        }
+        alert()->error('Gagal', 'Data gagal dihapus');
+        return redirect()->route('user.index');
     }
 
     public function resetPassword($id)
